Updated to current master status
